Fix for tax rounding and admin panel Orders reformatting

+++ src/Sylius/Component/Core/Model/OrderItemInterface.php
@@ -32,7 +32,7 @@ interface OrderItemInterface extends BaseOrderItemInterface, VersionedInterface
 
     public function setVariantName(?string $variantName): void;
 
-    public function getTaxTotal(): int;
+    public function getTaxTotal(): float;
 
     public function getDiscountedUnitPrice(): float;
